chore: rename vercel_crypto.py to crypto.py
