+++ ee/api/vercel/vercel_sso.py
@@ -16,7 +16,7 @@ from posthog.models.organization_integration import OrganizationIntegration
 from posthog.utils_cors import KNOWN_ORIGINS
 
 from ee.api.vercel.types import VercelUserClaims
-from ee.api.vercel.vercel_crypto import decrypt_payload, encrypt_payload, mark_token_used
+from ee.api.vercel.crypto import decrypt_payload, encrypt_payload, mark_token_used
 from ee.api.vercel.vercel_error_mixin import VercelErrorResponseMixin
 from ee.api.vercel.vercel_region_proxy_mixin import VercelRegionProxyMixin
 from ee.vercel.integration import SSOParams, VercelIntegration
